componentize the search input on pages
