+++ components/PageHeader.tsx
@@ -2,6 +2,10 @@
 import { useState, useEffect } from "react";
 import { useCardBorderHidden } from "./Pages/useCardBorderHidden";
 import { MediaContents } from "./Media";
+import { SearchSmall } from "./Icons/SearchSmall";
+import { MenuSmall } from "./Icons/MenuSmall";
+import { Separator } from "./Layout";
+import { CloseTiny } from "./Icons/CloseTiny";
 
 export const DesktopHeader = (props: { children: React.ReactNode }) => {
   let cardBorderHidden = useCardBorderHidden();
@@ -81,17 +85,39 @@ export const MobileHeader = (props: {
   controls?: React.ReactNode;
   actions: React.ReactNode;
 }) => {
+  let [state, setState] = useState<"search" | "default">("default");
   return (
     <MediaContents
       mobile={true}
-      className="mobilePageHeader flex gap-4 justify-between items-center px-1 mt-1 -mb-1 w-full "
+      className="mobilePageHeader container flex gap-4 justify-between items-center p-1 mt-1 -mb-1 w-full "
     >
-      {props.controls ? (
-        <div className="grow">{props.controls}</div>
+      {props.controls && state === "search" ? (
+        <>
+          {props.controls}
+          <Separator classname="h-6" />
+          <button
+            onClick={() => {
+              setState("default");
+            }}
+          >
+            <CloseTiny />
+          </button>
+        </>
       ) : (
-        <div className="spacer flex-1" />
+        <>
+          <MenuSmall /> Home
+          <div className="flex-1" />
+          <button
+            onClick={() => {
+              console.log("click");
+              setState("search");
+            }}
+          >
+            <SearchSmall />
+          </button>
+          <div className="flex flex-row-reverse! gap-1">{props.actions}</div>
+        </>
       )}
-      <div className="flex flex-row-reverse! gap-1">{props.actions}</div>
     </MediaContents>
   );
 };
